% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jggWrappers.R
\name{eq_geom_timeline}
\alias{eq_geom_timeline}
\title{eq_geom_timeline}
\usage{
eq_geom_timeline(df, xmin = 1, xmax = year(Sys.Date()), countries = "")
}
\arguments{
\item{df}{A data table containing NOAA Earthquake processed}

\item{xmin}{minimum year as numeric}

\item{xmax}{maximum year as numeric}

\item{countries}{Vector of countries to filter}
}
\value{
A ggplot2 graphical object displaying timeline of earthquakes data
}
\description{
A wrapper function to simplify use of geom_timeline
}
\examples{
\dontrun{
eq_geom_timeline(df)
eq_geom_timeline(df,xmin=2000,xmax=2015, countries=c("USA","IRAN"))
}
}
